Try to split out rich text components
