+++ spa/src/components/pageEditor/elementEditors/richText/RichTextElementEditor.tsx
@@ -1,6 +1,6 @@
 import { EditorState } from 'draft-js';
 import { useState } from 'react';
-import { RichTextEditor } from 'components/base';
+import { RichTextEditor } from 'components/base/RichTextEditor/RichTextEditor';
 import { useEditInterfaceContext } from 'components/pageEditor/editInterface/EditInterfaceContextProvider';
 import { editorStateToHtml, htmlToEditorState } from 'utilities/draftJs';
 import { Root } from './RichTextElementEditor.styled';
